TI - 94 - rebase develop

+++ smartimmo/src/main/java/com/gangdestrois/smartimmo/infrastructure/jpa/entity/PotentialProjectEntity.java
@@ -26,7 +26,7 @@ public class PotentialProjectEntity {
     public PotentialProjectEntity() {
     }
 
-    public PotentialProjectEntity(Integer id) {
+    public PotentialProjectEntity(Long id) {
         this.id = id;
     }
